Created Formatted API DOCS
Check the API endpoints when you can to start implementing them in the frontend @maufcost .
To summarize, you'll use /mint to mint & sell an NFT and then you'll be able to view all NFTs for sale using  /nftsforsale. After that when someone clicks on the NFT & tries to buy it, you'll have to use the data from /nftsforsale in order to send a /buy request to fulfill an order.